Addition of Segments in the Progress Bar
The color resources for ProgressColor and GradientColor have been reorganized in MainPage.xaml. Added a ScrollView to allow scrolling of the page content. Added a "Segment usage" section in MainPage.xaml to demonstrate the use of segments in the progress bar. Added new properties in ProgressBar.cs to manage segments: UseSegment, SegmentSpacing, SegmentCount, and ProgressSegment. Modified the DrawProgress method to include the logic for drawing segments if UseSegment is enabled. Added GenerateProgress and GenerateSegment methods to separate the drawing logic for progress and segments. The base color is now set as transparent when segments are used.